feat: Change to new version of plugin API
Change to the new version of the commit analyzer API. Now the parser is called
multiple times with only a single commit each time, and semantic-release takes
care of keeping track of the version. This drastically simplifies the code.

BREAKING CHANGE:
Instead of accepting an array of commits with the `commits` option, we now
accept a single commit with the `commit` option. Use that instead, and call
the plugin multiple times.